+++ README.md
@@ -13,7 +13,7 @@ Tested: `Ubuntu 5.9.0-050900rc6daily20200923-generic #202009222208 SMP Wed Sep 2
 ## First glance
 
 ```cpp
-#include "io_service.hpp"
+#include <liburing/io_service.hpp>
 
 int main() {
     // You first need an io_service instance
