+++ demo/http_client.cpp
@@ -11,7 +11,7 @@
 #include <cerrno>
 #include <fmt/format.h> // https://github.com/fmtlib/fmt
 
-#include "io_service.hpp"
+#include <liburing/io_service.hpp>
 
 task<> start_work(io_service& service, const char* hostname) {
     addrinfo hints = {
