+++ demo/threading.cpp
@@ -6,7 +6,7 @@
 
 #include <fmt/format.h>
 
-#include "io_service.hpp"
+#include <liburing/io_service.hpp>
 
 template <typename Fn>
 task<std::invoke_result_t<Fn>> invoke(io_service& service, Fn&& fn) noexcept(noexcept(fn())) {
