+++ include/liburing/task.hpp
@@ -5,7 +5,7 @@
 #include <array>
 #include <cassert>
 
-#include "stdlib_coroutine.hpp"
+#include <liburing/stdlib_coroutine.hpp>
 
 template <typename T, bool nothrow>
 struct task;
