+++ tests/delay_and_print.cpp
@@ -1,7 +1,7 @@
 #include <chrono>
 #include <fmt/chrono.h> // https://github.com/fmtlib/fmt
 
-#include "io_service.hpp"
+#include <liburing/io_service.hpp>
 
 int main() {
     io_service service;
